% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_splitting.R
\name{data_split}
\alias{data_split}
\title{Split data into training and testing sets}
\usage{
data_split(data, split_ratio = 0.7, seed = 123)
}
\arguments{
\item{data}{A data frame containing the dataset}

\item{split_ratio}{A numeric value representing the proportion of data to be used for training (default is 0.7)}

\item{seed}{An integer seed for reproducibility (default is 123)}
}
\value{
A list with two components:
\item{train}{A data frame containing the training set}
\item{test}{A data frame containing the testing set}
}
\description{
This function splits a dataset into training and testing sets.
A seed is used to ensure reproducibility.
}
\examples{
split_data <- data_split(mtcars, split_ratio = 0.7)

train_data <- split_data$train
test_data <- split_data$test

head(train_data)
head(test_data)
}
