% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.numeric.summary.R
\name{df.numeric.summary}
\alias{df.numeric.summary}
\title{Summary of Numeric Columns}
\usage{
df.numeric.summary(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data frame. Non-numeric columns will be removed}

\item{na.rm}{A logical indicating whether missing values should be removed}
}
\value{
This function returns a \code{data.frame} including columns:
 \itemize{
 \item id
 \item min
 \item q1
 \item mean
 \item median
 \item q3
 \item iqr
 \item stdev
 \item max
 \item skewness
 \item skew
}
}
\description{
Generate specific summaries of numeric columns in a data frame
}
\examples{
df.numeric.summary(iris)
df.numeric.summary(airquality, na.rm = FALSE)
}
\seealso{
\link[base]{summary}
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
