% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.plots.R
\name{do.plots}
\alias{do.plots}
\title{List of plots in a grid arrangement}
\usage{
\method{do}{plots}(df, funct, nr, ncols = 3)
}
\arguments{
\item{df}{A data frame.}

\item{funct}{function name. Tested for(\code{plot.histogram} or \code{plot.density}).}

\item{nr}{A range number of plots.}

\item{ncols}{Number of columns of the grid arrangement.}
}
\value{
List of plots in a grid arrangement
}
\description{
Grid of plots depending on the function (\code{plot.histogram} or \code{plot.density}) for a given df
}
\examples{
do.plots(iris, func=plot.density, nr=1:4, ncols=3)
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
