% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlhelper.R
\docType{package}
\name{mlhelper}
\alias{mlhelper}
\alias{mlhelper-package}
\title{mlhelper: Helper functions for EDA and ml}
\description{
Collection of EDA and helper df preprocess functions
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
