% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{cols}{number of columns}
}
\value{
This function returns the plots in a specified layout
}
\description{
ggplot objects can be passed in directly(...), or as a list of ggplot objects
}
\examples{
multiplot(p1, p2, cols=1)
}
\seealso{
\code{\link[base]{gridExtra}}
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
