% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.summary.R
\name{numeric.summary}
\alias{numeric.summary}
\title{Numeric Summaries}
\usage{
numeric.summary(x, na.rm)
}
\arguments{
\item{x}{a numeric vector containing the values to summarize.}

\item{na.rm}{A logical indicating whether missing values should be removed.}
}
\value{
This function returns a \code{data.frame} including columns:
\itemize{
 \item min
 \item q1
 \item mean
 \item median
 \item q3
 \item iqr
 \item stdev
 \item max
 \item skewness
 \item skew
}
}
\description{
Summarises numeric data and returns a data frame containing the basic summary values.
}
\examples{
numeric.summary(iris$Sepal.Length)
numeric.summary(airquality$Wind, na.rm = FALSE)
}
\seealso{
\code{\link[base]{summary}}
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
