% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.histogram.R
\name{plot.histogram}
\alias{plot.histogram}
\title{Histogram of a factor feature in a df}
\usage{
\method{plot}{histogram}(df, feature)
}
\arguments{
\item{df}{A data frame. Factor columns will be considered.}

\item{factor.v}{Factor feature}
}
\value{
Histogram of the feature
}
\description{
Plots a basic histogram for a factor feature in a given df
}
\examples{
plot.histogram(weather, "Weather")
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
