% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Plot feature importance for models}
\usage{
plot_feature_importance(model)
}
\arguments{
\item{model}{A trained model object}
}
\value{
A ggplot object showing feature importance
}
\description{
This function calculates and plots feature importance for models that support it.
}
\examples{
model <- train_model(mpg ~ ., mtcars, method = "rpart")
plot_feature_importance(model)
}
