% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_model.R
\name{predict_model}
\alias{predict_model}
\title{Make predictions using a trained model}
\usage{
predict_model(model, newdata)
}
\arguments{
\item{model}{A trained model object}

\item{newdata}{A data frame containing the new data}
}
\value{
A numeric vector of predictions. The length of the vector is equal to the number of rows in newdata.
}
\description{
This function makes predictions on new data using a trained model.
}
\examples{
model <- train_model(mpg ~ ., mtcars, method = "lm")
predictions <- predict_model(model, newdata = mtcars)

head(predictions)
}
