% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{RMSE  - Calcuates the Root Mean Squared Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{actual values taken from the original df}

\item{predicted}{predicted values taken from the model}
}
\value{
This function returns the calculated \code{rmse}
}
\description{
RMSE  - Calcuates the Root Mean Squared Error
}
\examples{
rmse(actual, predicted)
}
\author{
Eva Szin Takacs, \email{szin.takacs.eva@gmail.com}
}
