% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a machine learning model}
\usage{
train_model(formula, data, method = "lm")
}
\arguments{
\item{formula}{A formula specifying the model}

\item{data}{A data frame containing the dataset}

\item{method}{A character string specifying the model type (e.g., "lm", "rpart")}
}
\value{
A trained model object. The type of the returned object depends on the method used:
\item{lm}{If method is "lm", returns an object of class "lm"}
\item{rpart}{If method is "rpart", returns an object of class "rpart"}
}
\description{
This function trains a machine learning model using the provided formula and data.
}
\examples{

model <- train_model(mpg ~ ., mtcars, method = "lm")
summary(model)

model <- train_model(mpg ~ ., mtcars, method = "rpart")
summary(model)

}
