% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\docType{data}
\name{weather}
\alias{weather}
\title{Random Weather Data}
\format{A data frame of 7 rows and 3 columns
\describe{
 \item{Day}{Numeric values giving day of the week, 1 = Monday, 7 = Sunday}
 \item{Temp}{Numeric values giving temperature in degrees Celsius}
 \item{Weather}{Character values describing the weather on that day}
}}
\source{
Randomly generated data
}
\usage{
weather
}
\description{
A dataset containing randomly generated weather data.
}
\keyword{datasets}
